move function classes to analytics namespace
